Split python sources into logical chunks.
